requirements.txt
# Create virtual environment:-
# python -m venv .venv
# .\.venv\Scripts\activate.ps1

# After activating, install dependencies with:
# pip install -r requirements.txt

# Minimal runtime requirements for this app
Flask==3.1.2
python-dotenv==1.2.1
mysql-connector-python==9.5.0
joblib==1.5.3
numpy==2.4.0
pandas==2.3.3
scikit-learn==1.8.0
matplotlib==3.10.8
google-genai
